Added RecoveryState
Separated jump recovery from StandState
