+++ Assets/[BIMOS] Player Rig.prefab
@@ -5382,6 +5382,7 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 9d5b1d523de3f1a4584f4f93433eb62b, type: 2}
   - {fileID: 11400000, guid: 4a3867bfecfaa2a47be1caaaba19fa2b, type: 2}
   - {fileID: 11400000, guid: d3448344338213645bbb346627ab41f5, type: 2}
+  - {fileID: 11400000, guid: bd24e3d61c8e03a458c0bebb9a8abd2b, type: 2}
   InitialState: {fileID: 11400000, guid: 6b8c9b828fb013e4382879494b093b69, type: 2}
 --- !u!114 &2259126866414753448
 MonoBehaviour:
