+++ Runtime/Rig/Movement/Jumping/States/RecoverState.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd24e3d61c8e03a458c0bebb9a8abd2b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
