+++ Runtime/Rig/Movement/Jumping/States/RecoverState.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+
+namespace KadenZombie8.BIMOS.Rig.Movement
+{
+    public class RecoverState : StandState
+    {
+        protected override void Update()
+        {
+            var fullHeight = Crouching.StandingLegHeight - Crouching.CrouchingLegHeight;
+            Crouching.TargetLegHeight += Crouching.VirtualCrouching.CrouchSpeed * fullHeight * Time.deltaTime;
+
+            if (Crouching.TargetLegHeight > Crouching.StandingLegHeight)
+                StateMachine.ChangeState<StandState>();
+
+            if (Crouching.VirtualCrouching.IsCrouchChanging)
+                StateMachine.ChangeState<StandState>();
+        }
+    }
+}
